Add position type check to layout metrics == operator (#42020)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/42020

I added position type in D51412428 (https://github.com/facebook/react-native/pull/41819). I didn't notice this == override which makes it so position type in layout metrics will not be updated if it changes.

To use this cpp 20 feature we needed to change a few buck files which is also done here

Changelog: [Internal]

Reviewed By: NickGerleman

Differential Revision: D52339890

fbshipit-source-id: e77ee092477dbf786e4a72e6a33138ccbc450645