Change RN default position type to relative (#42062)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/42062

We have had relative as the default for a few big frameworks/apps right now (Fabric FB iOS, Fabric FB Android, OSS, Litho, CK) and have not run into issues. Seems it is safe to pull the trigger here and put everything on relative 🎉

This also fixes a test that relied on this default, changes the layout metrics default, and removes the gating plumbing that was in place earlier.

Lastly, a few animation tests start failing after this change. Seems that there is an animation bug with relative trees that would have existed already, so this is merely discovering that that bug exists, not causing any extra issues. Since that test is a set of random trees with random props it is very hard to debug and I am just adding skips to the failing ones.

Changelog: [Internal]

Reviewed By: NickGerleman

Differential Revision: D52137858

fbshipit-source-id: 6856bc608b8211c868c9ee81fc92e005ec3d2faa