Add gtests for order index logic (#42022)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/42022

Some tests to ensure that the order of nodes is correct as defiend by order index that is derived from zIndex + positioning. These tests do not ensure that the native platform then goes and lays them out correctly. That will be done later with e2e tests on catalyst

Changelog: [Internal]

Reviewed By: NickGerleman

Differential Revision: D52295063

fbshipit-source-id: 8ae29fc50ad65db8e5c0ab28a132546d8489dffe