Reintroduce OldArch Integration Tests (#42262)
Summary:
Yesterday we landed a change that removed tests for the Old Architecture for RNTester.
That was the right call as there are no build differences in RNTester between the two architectures. But we do have runtime differences, and we had an integration test running on RNTester that we deleted with the previous PR.
This change restores that test, adding this only new job to run that test

## Changelog:
[Internal] - Add back an old arch integration test

Pull Request resolved: https://github.com/facebook/react-native/pull/42262

Test Plan: CircleCI is green

Reviewed By: cortinico

Differential Revision: D52730661

Pulled By: cipolleschi

fbshipit-source-id: 10fbc2540abeebc72f635451f6f650827cf20041