Fabric Automatic Interop for Android (#42294)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/42294

This changes enables the Fabric Intrerop Layer automatically for all the users.

Practically I'm removing the fallback `ComponentDescriptor` inside the `ComponentDescriptorRegistry` so that if a component hasn't been automatically registered by a user, instead of showing the UnimplementedView, it loads a `UnstableLegacyViewManagerAutomaticComponentDescriptor`.

This ComponentDescriptor is built starting from the legacy component name, and responds with correct `ComponentName` and `ComponentHandle` (similarly to the `UnstableLegacyViewManagerInteropComponentDescriptor` but without using C++ templates).

Changelog:
[Internal] [Changed] - Fabric Automatic Interop for Android

Reviewed By: sammy-SC

Differential Revision: D52663244

fbshipit-source-id: f466486c638bb1362ef59128cd69bb9731bb9739